+++ CHANGELOG.md
@@ -17,6 +17,16 @@ __Reference__:
 
 # CHANGELOG
 
+## [0.2.0]
+
+### Added
+
+* Example on how through loop through nested dictionaries.
+
+### Changed
+
+* Print statements to use Print(f"")
+
 ## [0.1.0]
 
 ### Added